Prevent overflow causing visual issue
In some cases the content of the page can be longer than the page
height. The GOV.UK template background is different than the dark grey
background we use, so when this happens everything below the fold will
have a different background colour, which looks wrong.

This commit updates the stylesheet so that the html and body elements
will have a minimum height of 100vh (i.e.e the size of the viewport).
This means that:

- if the content is smaller than the viewport, the elements with the
  dark background will expand to cover the whole viewport
- if the content overflows the viewport, the elements with the dark
  background will expand enough to fit the content.